refactor: change paths to utility
