+++ src/features/input/elements/Button/Button.js
@@ -1,5 +1,5 @@
 import data from './Button.data.json'
-import createCustomizedElement from '../../utils/createCustomizedElement'
+import createCustomizedElement from '../../../shared/utils/createCustomizedElement'
 
 const Button = createCustomizedElement(data.name, data.attrs, data.children)
