+++ src/features/input/elements/Form/Form.js
@@ -1,4 +1,4 @@
-import createCustomizedElement from '../../utils/createCustomizedElement'
+import createCustomizedElement from '../../../shared/utils/createCustomizedElement'
 import data from './Form.data.json'
 
 const Form = createCustomizedElement(data.name, data.attrs, data.children)
