+++ src/features/input/elements/Input/Input.js
@@ -1,5 +1,5 @@
 import data from './Input.data.json'
-import createCustomizedElement from '../../utils/createCustomizedElement'
+import createCustomizedElement from '../../../shared/utils/createCustomizedElement'
 
 const Input = createCustomizedElement(data.name, data.attrs, data.children)
